Expose Encoding randomHex
