+++ .changeset/random-hex-encoding.md
@@ -1,0 +1,5 @@
+---
+"effect": patch
+---
+
+Add `Encoding.randomHex` for generating random lowercase hexadecimal strings.
